requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.5
numpy==1.26.4
packaging==25.0
pillow==11.3.0
scikit-image==0.25.2
scikit-surprise==1.1.4
scipy==1.16.0
tifffile==2025.6.11
Werkzeug==3.1.3
